Add files via upload
Jeff's requested changes completed: Can we get the slider itself to announce a name that aligns with the header for it? Right now, it says "Cook time, heading level 3" then "Range. 4 to 120 minutes." followed by "Slider 4" I think it would be better to have it announce "Cook time range, 4 to 120 minutes." and "Cook time sider"